Refactor: move GridViewHolder into GridAdapter
